+++ src/Connection.php
@@ -7,7 +7,9 @@ namespace Yiisoft\Db\Pgsql;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoConnection;
 use Yiisoft\Db\Driver\Pdo\PdoCommandInterface;
 use Yiisoft\Db\Exception\InvalidArgumentException;
+use Yiisoft\Db\Pgsql\Column\ColumnFactory;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 use Yiisoft\Db\Schema\Quoter;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
@@ -44,6 +46,11 @@ final class Connection extends AbstractPdoConnection
         return new Transaction($this);
     }
 
+    public function getColumnFactory(): ColumnFactoryInterface
+    {
+        return new ColumnFactory();
+    }
+
     public function getLastInsertID(?string $sequenceName = null): string
     {
         if ($sequenceName === null) {
@@ -55,28 +62,16 @@ final class Connection extends AbstractPdoConnection
 
     public function getQueryBuilder(): QueryBuilderInterface
     {
-        return $this->queryBuilder ??= new QueryBuilder(
-            $this->getQuoter(),
-            $this->getSchema(),
-            $this->getServerInfo(),
-        );
+        return $this->queryBuilder ??= new QueryBuilder($this);
     }
 
     public function getQuoter(): QuoterInterface
     {
-        if ($this->quoter === null) {
-            $this->quoter = new Quoter('"', '"', $this->getTablePrefix());
-        }
-
-        return $this->quoter;
+        return $this->quoter ??= new Quoter('"', '"', $this->getTablePrefix());
     }
 
     public function getSchema(): SchemaInterface
     {
-        if ($this->schema === null) {
-            $this->schema = new Schema($this, $this->schemaCache);
-        }
-
-        return $this->schema;
+        return $this->schema ??= new Schema($this, $this->schemaCache);
     }
 }
