+++ src/QueryBuilder.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql;
 
-use Yiisoft\Db\Connection\ServerInfoInterface;
+use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Pgsql\Column\ColumnDefinitionBuilder;
 use Yiisoft\Db\QueryBuilder\AbstractQueryBuilder;
-use Yiisoft\Db\Schema\QuoterInterface;
-use Yiisoft\Db\Schema\SchemaInterface;
 
 use function bin2hex;
 
@@ -17,12 +15,13 @@ use function bin2hex;
  */
 final class QueryBuilder extends AbstractQueryBuilder
 {
-    public function __construct(QuoterInterface $quoter, SchemaInterface $schema, ServerInfoInterface $serverInfo)
+    public function __construct(ConnectionInterface $db)
     {
+        $quoter = $db->getQuoter();
+        $schema = $db->getSchema();
+
         parent::__construct(
-            $quoter,
-            $schema,
-            $serverInfo,
+            $db,
             new DDLQueryBuilder($this, $quoter, $schema),
             new DMLQueryBuilder($this, $quoter, $schema),
             new DQLQueryBuilder($this, $quoter),
